Endring i utbetaling: trekk ut logikk som genererer tidslinje pr person og ytelsetype
